# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: WeCratfs API
  version: 1.0.0
  description: JSON-only v1 contract for the WeCratfs marketplace.
servers:
  - url: /
paths:
  /api/v1/products:
    get:
      summary: List approved products
      parameters:
        - {name: category, in: query, schema: {type: string}}
      responses:
        '200': {description: Product list}
...
  /api/v1/orders:
    get:
      summary: List the current customer's orders
      security: [{cookieSession: []}]
      responses: {'200': {description: Orders}}
...
components:
  securitySchemes:
    cookieSession: {type: apiKey, in: cookie, name: wecratfs_session}
    csrfHeader: {type: apiKey, in: header, name: X-CSRF-Token}
...